requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
websockets>=11.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0

# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
psutil>=5.9.0

# Machine Learning and NLP
autogen~=0.8.1
openai>=1.0.0
spacy>=3.7.0
transformers>=4.35.0
torch>=2.1.0

# HTTP and API clients
httpx>=0.25.0
requests>=2.31.0
aiohttp>=3.9.0

# Data collection and web scraping
selenium>=4.11.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
feedparser>=6.0.10
apify-client>=1.3.0
chardet>=5.0.0

# Google API clients
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0

# Email
aiosmtplib>=3.0.0

# Scheduling and monitoring
apscheduler>=3.10.0
schedule>=1.2.0
python-json-logger>=2.0.0
structlog>=23.2.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
tqdm>=4.65.0
tabulate>=0.9.0
schedule>=1.2.0
# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
mypy>=1.7.0
types-requests>=2.31.0
types-python-dateutil>=2.8.19

# Additional utilities
hf_xet




